requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.2.2
